refactor: extract shared styles and copy logic to reusable modules
